backend/requirements.txt
Flask>=3.0,<4
flask-cors>=4.0,<5
firebase-admin>=6.5,<7
google-cloud-firestore>=2.16,<3
pydantic>=2.8,<3
python-dotenv>=1.0,<2